refactor(tooltip): refactor component code, tests and docs
