--- packages/web/src/Components/Tooltip/ITooltipProps.ts
@@ -1,14 +1,0 @@
-export { TooltipProps as ITooltipProps } from "@material-ui/core/Tooltip";
-
-export type IPlacement = "bottom-end"
-  | "bottom-start"
-  | "bottom"
-  | "left-end"
-  | "left-start"
-  | "left"
-  | "right-end"
-  | "right-start"
-  | "right"
-  | "top-end"
-  | "top-start"
-  | "top";
