+++ packages/web/src/Components/Tooltip/Tooltip.props.ts
@@ -1,0 +1,3 @@
+import { TooltipProps } from "@material-ui/core/Tooltip";
+
+export type ITooltipProps = TooltipProps;
